requirements.txt
faiss-cpu
neo4j
ollama
openai
sentence-transformers
tiktoken
rank-bm25
rapidfuzz
streamlit
hazm
python-dotenv
